+++ src/main/java/com/example/backend/Service/PalierService.java
@@ -3,6 +3,7 @@ package com.example.backend.Service;
 import com.example.backend.DTO.PalierDto;
 import com.example.backend.Exception.ResourceNotFoundException;
 import com.example.backend.Model.Palier;
+import com.example.backend.Model.Region;
 import com.example.backend.Model.Utilisateur;
 import com.example.backend.Repository.PalierRepository;
 
@@ -18,14 +19,23 @@ import java.util.UUID;
 public class PalierService {
 
     private final PalierRepository palierRepository;
-    private final UtilisateurRepository utilisateurRepository;
+
 
     public PalierService(PalierRepository palierRepository, UtilisateurRepository utilisateurRepository) {
         this.palierRepository = palierRepository;
-        this.utilisateurRepository = utilisateurRepository;
     }
 
     public Palier createPalier(Palier palier) {
+
+        if (palierRepository.existsByCode(palier.getCode())) {
+            throw new IllegalArgumentException("Un palier avec ce code existe déjà : " + palier.getCode());
+        }
+
+
+        if (palierRepository.existsByLabel(palier.getLabel())) {
+            throw new IllegalArgumentException("Un Palier avec ce label existe déjà : " + palier.getLabel());
+        }
+
         return palierRepository.save(palier);
 
     }
@@ -44,8 +54,11 @@ public class PalierService {
             return palierRepository.save(palier);
         }).orElseThrow(() -> new ResourceNotFoundException("Palier non trouvé avec l'id : " + id));
     }
-
     public void deletePalier(UUID id) {
+
+        if (!palierRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Impossible de supprimer : Palier introuvable avec l'ID : " + id);
+        }
         palierRepository.deleteById(id);
     }
 }
