+++ src/main/java/com/example/backend/Service/SegmentService.java
@@ -25,13 +25,21 @@ public class SegmentService {
     }
 
     public Segment createSegment(Segment segment) {
+
+        if (segmentRepository.existsByCode(segment.getCode())) {
+            throw new IllegalArgumentException("Un segment avec ce code existe déjà : " + segment.getCode());
+        }
+        if (segmentRepository.existsByLabel(segment.getLabel())) {
+            throw new IllegalArgumentException("Un segment avec ce label existe déjà : " + segment.getLabel());
+        }
+
         if (segment.getMarche() == null || segment.getMarche() .getId() == null) {
-            throw new ResourceNotFoundException("marche must be provided with a valid ID");
+            throw new ResourceNotFoundException("Le marché doit être fourni avec un ID valide");
         }
         else {
             UUID marcheId =  segment.getMarche() .getId();
             Marche marche = marcheRepository.findById(marcheId)
-                .orElseThrow(() -> new ResourceNotFoundException("Zone not found with ID: " + marcheId));
+                .orElseThrow(() -> new ResourceNotFoundException("Marché introuvable avec l'ID: " + marcheId));
              segment.setMarche(marche);
              return segmentRepository.save(segment);
     }}
@@ -49,7 +57,7 @@ public class SegmentService {
             segment.setLabel(segmentDetails.getLabel());
             UUID marcheId =  segmentDetails.getMarche() .getId();
             Marche marche = marcheRepository.findById(marcheId)
-                    .orElseThrow(() -> new ResourceNotFoundException("Zone not found with ID: " + marcheId));
+                    .orElseThrow(() -> new ResourceNotFoundException("Marché introuvable avec l'ID " + marcheId));
             segmentDetails.setMarche(marche);
             segment.setMarche(segmentDetails.getMarche());
             return segmentRepository.save(segment);
@@ -57,6 +65,10 @@ public class SegmentService {
     }
 
     public void deleteSegment(UUID id) {
+
+        if (!segmentRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Impossible de supprimer : Segment introuvable avec l'ID : " + id);
+        }
         segmentRepository.deleteById(id);
     }
 }
