+++ src/main/java/com/example/backend/Service/ZoneService.java
@@ -24,13 +24,22 @@ public class ZoneService {
     }
 
     public Zone createZone(Zone zone) {
+        if (zoneRepository.existsByCode(zone.getCode())) {
+            throw new IllegalArgumentException("Une zone avec ce code existe déjà : " + zone.getCode());
+        }
+
+
+        if (zoneRepository.existsByLabel(zone.getLabel())) {
+            throw new IllegalArgumentException("Une zone avec ce libellé existe déjà : " + zone.getLabel());
+        }
+
         if(zone.getRegion() == null || zone.getRegion().getId()==null) {
-            throw new ResourceNotFoundException("Region not found");
+            throw new ResourceNotFoundException("La région doit être fournie avec un ID valide");
         }
         else {
             UUID regionId = zone.getRegion().getId();
             Region region= regionRepository.findById(regionId)
-                    .orElseThrow(() -> new ResourceNotFoundException("Region not found with ID: " + regionId));
+                    .orElseThrow(() -> new ResourceNotFoundException("Région introuvable avec l'ID: " + regionId));
 
             zone.setRegion(region);
             return  zoneRepository.save(zone);
@@ -55,15 +64,18 @@ public class ZoneService {
 
             UUID regionId = zoneDetails.getRegion().getId();
             Region region= regionRepository.findById(regionId)
-                    .orElseThrow(() -> new ResourceNotFoundException("Region not found with ID: " + regionId));
+                    .orElseThrow(() -> new ResourceNotFoundException("Région introuvable avec l'ID " + regionId));
 
             zoneDetails.setRegion(region);
             zone.setRegion(zoneDetails.getRegion());
             return zoneRepository.save(zone);
-        }).orElseThrow(() -> new ResourceNotFoundException("Zone non trouvée avec l'id : " + id));
+        }).orElseThrow(() -> new ResourceNotFoundException("Zone introuvable avec l'ID : " + id));
     }
 
     public void deleteZone(UUID id) {
+        if (!zoneRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Impossible de supprimer : Zone introuvable avec l'ID : " + id);
+        }
         zoneRepository.deleteById(id);
     }
 }
